Factor out navigation params
